Create TimesSquareGitHubPagePanel component
The idea is that each type of page can have its own panel component that
is slotted into the sidebar. This component holds page metadata (title,
description, etc) as well as the page's parameters.

This effectively replaces the former TimesSquarePage component; note
that the notebook iframe was already separated out from TimesSquarePage.
