+++ src/components/TimesSquareApp/TimesSquareApp.js
@@ -20,11 +20,11 @@ const StyledLayout = styled.div`
   }
 `;
 
-export default function TimesSquareApp({ children, pageNav }) {
+export default function TimesSquareApp({ children, pageNav, pagePanel }) {
   console.log('Running TimesSquareApp');
   return (
     <StyledLayout>
-      <Sidebar pageNav={pageNav} />
+      <Sidebar pageNav={pageNav} pagePanel={pagePanel} />
       <main>{children}</main>
     </StyledLayout>
   );
