+++ src/components/TimesSquareGitHubPagePanel/index.js
@@ -1,0 +1,2 @@
+export * from './TimesSquareGitHubPagePanel';
+export { default } from './TimesSquareGitHubPagePanel';
