--- src/components/TimesSquarePage/index.js
@@ -1,2 +1,0 @@
-export * from './TimesSquarePage';
-export { default } from './TimesSquarePage';
