+++ src/hooks/useTimesSquarePage.js
@@ -8,6 +8,9 @@ function useTimesSquarePage(pageUrl) {
   return {
     error: error,
     loading: !error && !data,
+    title: data ? data.title : null,
+    parameters: data ? data.parameters : null,
+    description: data ? data.description : null,
     htmlUrl: data ? data.html_url : null,
     htmlStatusUrl: data ? data.html_status_url : null,
   };
